+++ app/Http/Controllers/UserController.php
@@ -15,6 +15,7 @@ class UserController extends Controller
     }
 
     public function index() {
+        $this->authorize('create.user');
         $user   =   User::all();
 
         return view('user.index' , ['user'=>$user]);
@@ -22,12 +23,15 @@ class UserController extends Controller
 
     public function create()
     {
+        $this->authorize('create.user');
         return view('user.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create.user');
         $user =  new User();
+        $user->unamea = $request->uname;
         $user->name = $request->name;
         $user->email = $request->email;
         $user->nama_ketua = $request->nama_ketua;
@@ -38,12 +42,14 @@ class UserController extends Controller
     }
 
     public function updateview($id) {
+        $this->authorize('create.user');
         $user = User::find($id);
 
         return view('user.edit', ['user'=>$user]);
     }
 
     public function update(Request $request, $id) {
+        $this->authorize('create.user');
         $user  = User::find($id);
         $user->name = $request->name;
         $user->email = $request->email;
@@ -58,6 +64,7 @@ class UserController extends Controller
     }
 
     public function delete($id) {
+        $this->authorize('create.user');
         User::find($id)->delete();
 
         return back();
